refactor: 주문에서 차감 로직 사용하지 않는 부분 수정 및 바깥쪽 dto 사용 로직 수정

+++ apps/commerce-api/src/main/java/com/loopers/domain/point/Point.java
@@ -12,30 +12,30 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Point {
 
-  private int amount;
+  private long amount;
 
-  public Point(int amount) {
+  public Point(long amount) {
     validate(amount);
     this.amount = amount;
   }
 
-  public Point add(int value) {
+  public Point add(long value) {
     if(value <= 0) throw new CoreException(ErrorType.BAD_REQUEST, "충전 금액이 부족합니다.");
     return new Point(this.amount + value);
   }
 
-  public Point subtract(int value) {
+  public Point subtract(long value) {
     if (this.amount < value) {
       throw new CoreException(ErrorType.BAD_REQUEST, "포인트 잔액이 부족합니다.");
     }
     return new Point(this.amount - value);
   }
 
-  private void validate(int amount) {
+  private void validate(long amount) {
     if (amount < 0) throw new CoreException(ErrorType.BAD_REQUEST);
   }
 
-  public int getAmount() {
+  public long getAmount() {
     return amount;
   }
 }
